Use widget and detail for cancer biomarkers
